requirements.txt
python-xlib>=0.31
pyautogui>=0.9.53
opencv-python>=4.7.0
numpy>=1.23.0
Pillow>=9.4.0
pytesseract>=0.3.10
openai>=0.27.0
